+++ Sources/SwiftletData/Model/Link.swift
